Fix 'flashing' issue
Fixed issue which caused Hello Triangle application to somewhat randomly flash
the image we wished to draw. The reason this was happening was because the
logic for backing textures with the currentDrawable in 'queueSubmit' was faulty.
We were only backing textures that were 'nil' with the current drawable texture.
Because command buffers are re-used in Fever's API, this was only true the first
time around. So color attachments referencing drawables were only getting the
correct drawable image the first time around, after that point they were keeping
a reference to the old drawable image - and drawing to that. This drawable
happened to be the current drawable every x frames, and so the image flashed
between black and the desired image seemingly at random.